requirements.txt
flask
python-dotenv
openai
langchain
langchain-openai
pypdf
unstructured[pdf]
pymupdf
faiss-cpu